requirements.txt
fastapi
uvicorn
flake8
python-dotenv